Fixed embed objects
